Rename Xgit.Repository.WorkingTree.ParseIndexFile.from_iodevice/1 to Xgit.DirCache.from_iodevice/1.
