+++ mix.exs
@@ -77,7 +77,6 @@ defmodule Xgit.MixProject do
           "Xgit.Tree",
           "Xgit.Tree.Entry",
           "Xgit.Repository.WorkingTree",
-          "Xgit.Repository.WorkingTree.ParseIndexFile",
           "Xgit.Repository.WorkingTree.WriteIndexFile"
         ]
       ]
